refactor(Envars): Remove excess return methods.
Removed return methods. All the setting lists are global and can be accessed directly from the module. There was no need for the extra layer.

Removed uses of the return methods from other modules.

Plan on moving settings into a class to put a layer of abstraction between them and the user so they don't get erroneously changed.