feature(Mongo):Enable asynchronous database operations
New feature allows for slower database operations, such as when interacting with a remote DB, to take place asynchronously.

The major benefit of this is that now when a user initiates an action that takes time, a loading screen is able to populate. This is because the slow actions take place on a separate thread - leaving Py_CUI to continue cycling through its drawing actions.

This can be expanded for several other functions.

- Fetching a query: The call from NCBI is not asynchronous, as it has not been slow for such small dev-queries. But we can make it asynchronous if larger queries take long.

- Connecting to the database: Any action that requires the database first checks if it has been started, and if not it starts it.